Update patch set 8
Patch Set 8: Code-Review-1

-1 for visibility
I expect that nssmf is just a user of this password if that's true then where is the owner of this password?

Is it SO and this config file: kubernetes/so/resources/config/overrides/override.yaml ?

If so then how to turn this: $2a$10$Fh9ffgPw2vnmsghsRD3ZauBL1aKXebigbq3BB1RPWtE62UDILsjke into password1$ or the other way around?

Patch-set: 8
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1