Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> @Krishna
> any hints how to force those components to log to stdout?

I once had a look at it before. Seems the logback file that is being mounted from the charts are not properly fed to the application. The application uses an internal logback file within the image. I will check to see if it could be overridden by this logback or we need remove the logback from the image and try.

Patch-set: 2
Reviewer: Gerrit User 2671 <2671@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>