[CCSDK] Fix error logs due to empty URL
The gating tests for Honolulu have reported flooding of error messages
in SDNC because the ricLink in values.yaml was not set.
Since there are no RICs provisioned in the default OOM instllation,
the ricLink has been removed.
The user needs to add the information about RICs separately in
application_configuration.json file when RICs are available.

Change-Id: I8ef67c27ed95432fb90de88b2976f39068fa7be0
Issue-ID: CCSDK-3227
Signed-off-by: RehanRaza <muhammad.rehan.raza@est.tech>