Update patch set 3
Patch Set 3: Code-Review-1

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > images have been given far after M3.
> > Will wait for honolulu branching to merge
> 
> Hi,
> We have already created honolulu branch for UUI, and also we have changed the codes and self released new version after M3. What should we do?

The problem is that this is the first container image that you are releasing in Honolulu so I expect it's not a minor bugfix but a major update compared to Guilin. Deadline for creating review with new containers was at M3 which was 4 weeks ago which means that you are 4 weeks after the deadline.

Just to be clear, we are not against this patch. If it works then as soon as we branch out Honoulu (probably on Monday) we'll merge it to the master branch and it's just going to become a part of Istanbul release

Patch-set: 3
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=-1