Update patch set 3
Patch Set 3:

This suggested enhanced change would be quite substantial, and unfair to ask for it to be merged for Honolulu. 

(This entire way of configuring RICs via OOM needs a fundamental rethink anyway) 

So I suggest abandoning this change for now - to be fixed in Istanbul.

Patch-set: 3
Reviewer: Gerrit User 5519 <5519@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>