Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> This suggested enhanced change would be quite substantial, and unfair to ask for it to be merged for Honolulu. 
> 

Well I considered it as a fix to our helm charts so I've been happy to take it even after RC0.

> (This entire way of configuring RICs via OOM needs a fundamental rethink anyway) 
> 
> So I suggest abandoning this change for now - to be fixed in Istanbul.

My understanding was that currently if ricLink is empty then SDNC gets somehow broken? Isn't it the case?

If it is then I can take the patch as is as a quick fix or if you have issues with helm I can help you with the suggested solution

Patch-set: 3
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>