Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > This suggested enhanced change would be quite substantial, and unfair to ask for it to be 
> > > merged for Honolulu. 
> > > 
> > 
> > Well I considered it as a fix to our helm charts so I've been happy to take it even after RC0.
> > 
> > > (This entire way of configuring RICs via OOM needs a fundamental rethink anyway) 
> > > 
> > > So I suggest abandoning this change for now - to be fixed in Istanbul.
> > 
> > My understanding was that currently if ricLink is empty then SDNC gets somehow broken? Isn't it
> > the case?
> > 
> > If it is then I can take the patch as is as a quick fix or if you have issues with helm I can help 
> > you with the suggested solution
> 
> That's exactly right - without this fix, the RIC ip ends up mangled, and the a1 policy manager 
> keeps polling SDNC with a missing RIC ... and SDNC keeps logging that message over and over and
> over, to the point where the logs roll and we miss real log messages.  This is why it took about
> 2 weeks to diagnose the issue in our version bump change.

A fix has been submitted in CCSDK to check for such a mis-configuration 
When a mis-configuration is found the RIC configuration is rejected - With a single ERROR entry in the log, then there will be no further log entry. 
Since the RIC configuration is rejected there is not periodic attempt to sync - so no repeated log error entry. 
The log flooding is then prevented. 

(Of course "normal" operation will still create "normal" INFO log entries - and this is probably too verbose - but we will address that in Istanbul 

Thanks for offering to merge a fix to the mis-configuration mentioned - even after RC0 - so we will rush to get this fix available ASAP.

Patch-set: 3
Reviewer: Gerrit User 5519 <5519@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>