Update patch set 13
Patch Set 13:

> Patch Set 13:
> 
> Can we merge the code to unlock SO-3590? thanks

As noted earlier we are waiting for a successful gating on https://gerrit.onap.org/r/c/oom/+/119309/24 to be sure that this change is really fixing the issue. Without this it would be just a blind merge.

The patch looks fine. There is no objections to it. We are just waiting for patch above that really makes use of it just to be sure that it works

Patch-set: 13
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>