Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> @Krzysztof; this change maybe arrived after M3 but typically we can still update the image version until RC2. In the new release cadence, can the project only do it at the M3? If it is the case then my apologies I have not yet captured this change. We should communicate to the PTLS on Monday 3/29 because I am not sure if all of them are aware of this change versus what we did before.

@Catherine
It's perfectly fine to provide bugfixes after M3. It's human to err and make bugs but in my opinion this is not a bugfix. It's a first update of this container that we are seeing in this release so I expect it to be a functional change not just a bugfix update of the container

Patch-set: 4
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>