Update patch set 14
Patch Set 14: Code-Review+1

https://logs.onap.org/onap-integration/gating//118510-14/index.html
note the pnf-registrate artifacts indicate a failure but as pnf-registrate is run as a job, a retry is done in case of errors (5 attempts). And if the status is green it means that a retry was successful.

Patch-set: 14
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1