Update patch set 29
Patch Set 29:

> Patch Set 29:
> 
> PS28 still doesn't work: https://gitlab.com/Orange-OpenSource/lfn/onap/xtesting-onap/-/pipelines/278438561
> You should get full results posted shortly but I already see that basic_cnf, basic_vnf and pnf_registrate were failed

For basic_cnf, it looks like some sort of timing issue in SO - they are rejecting our asynchronous response with an error indicating they are not expecting a response for that transaction.

For basic_vm, we are getting a 401 Unauthorized response from SO. This is the same error we got before my patch to set the SO credential variables. However, it doesn’t make sense that we’d get a 401 in this case but not for basic_cnf.

Let’s see if we have better luck with my last change set.

Patch-set: 29
Reviewer: Gerrit User 87 <87@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>