Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> I just see logs before tomcat starts on some:
> https://logs.onap.org/onap-integration/gating//113414-5/infrastructure-healthcheck/k8s-teardown/kubernetes-status/pod-onap-oof-cmso-topology-5c44ff6797-qnp6z-oof-cmso-topology-logs.html
> https://logs.onap.org/onap-integration/gating//113414-5/infrastructure-healthcheck/k8s-teardown/kubernetes-status/pod-onap-oof-cmso-ticketmgt-5b78895695-vkwsz-oof-cmso-ticketmgt-logs.html
> 
> but not on others so I guess it's OK.
> 
> I still  have a question:
> what is mso simulator in OOF CMSO service? https://logs.onap.org/onap-integration/gating//113414-5/infrastructure-healthcheck/k8s-teardown/kubernetes-status/pod-onap-oof-cmso-service-67d646d869-wgq8m.html
> 
> do we really need it?

Yes. I checked with one of the developer who worked in CMSO. CMSO was originally introduced as part of the change management use case, CMSO was planned to be integrated with SO. But it seems that it never took off. So, for the testing they used a simulator and included it in the helm charts too, for running some e2e testing.

Removing that side car container should not create any problem. Can you create a jira for me? I will take care of it.

Patch-set: 5
Reviewer: Gerrit User 2671 <2671@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>