Update patch set 9
Patch Set 9:

I have found the intension of the code here.
This code is to access the so's api handler endpoint through the server-infra-creds.
The way its done is little complex, but the intension is to have the access to the new service added in the api-handler accessed from the nssmf adapter.
you can see the override.yml has the cred details of the mso.infra

Patch-set: 9
Reviewer: Gerrit User 897 <897@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>