Update patch set 3
Patch Set 3:

> Patch Set 3: Code-Review+1
> 
> LGTM but I've a stupid question. Does it work? xD
> I'm asking because I'm unsure how k8s is going to behave when we try to configure ingress before this job succeeds and secret gets created?
> I'm also thinking if we shouldn't first create this secret using helm and later only update it with right values just to be sure that it doesn't stay as a leftover after we run helm uninstall
> 
> @Sylvain
> What do you think?

for Ingress, as long as the secret is not there, Ingress will use autogenerated certificate of ingress controller.
for leftover part, I also thought about it and if we helm undeploy __without__ ns deletion after, there will be leftover.
But K8S doesn't like when you update a resource that doesn't belongs to "you" (so kubectl apply on resources created by helm triggers warning) so I guess we could do via helm hook.
But I'd prefer to merge this and then later see how we can improve...

Patch-set: 3