Update patch set 5
Patch Set 5: Code-Review+1

Dears - let`s cool down a little.
There is already a PS5, which refers to an image, which does not contain Helm v2.

The next step to to validate that image in gating, if that one works well, we`ll merge it.
The change will be introduced in Istanbul, and we expect MultiCloud-K8S to update to Helm v3 in this timeframe. 

If MultiCloud-K8S doesn`t update in Istanbul time-frame, then well - then we go back to the discussion, if ONAP is one project, or a set of loosely coupled projects.

Adding +1 on PS5.

Patch-set: 5
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1