Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> going back to guilin with this? Is this a part of maintenance release? Does it make sense?
> -1 for visibility

this is for a bug fix that was recently reported in Guilin, long after the maintenance release was planned: https://jira.onap.org/browse/POLICY-3177.  Weren't sure how to handle it, since it's in OOM.

however, it's OK with me if we do not include it in Guilin.  In that case, would you please update the JIRA and explain to fabian that the fix is included in Honolulu and going forward?  If he insists that it needs to be in Guilin, too, then we can continue with this cherry pick.

Patch-set: 1
Reviewer: Gerrit User 811 <811@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>