Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > Patch Set 8:
> > 
> > > Patch Set 8: Code-Review-1
> > > 
> > > What's the reason for removing bashism within bash scripts (see shebang)? Especially if they introduce another security risk (eval)?
> > 
> > see first change in the relation chain for explanations
> > https://gerrit.onap.org/r/c/oom/+/118922/5
> > shebangs are migrated at the end of the relation chain when all bashisms have been fixed
> > 
> > Only one script (db-init.sh) might present an issue here.
> > db-init.sh retrieves some DB config parameters from helm and uses them as input for eval.
> > It's very likely these values are constrained too and it presents no risk.
> > But I need a helm specialist's opinion to confirm that.
> > 
> > The 2 others scripts are safe since all possible eval inputs are listed inside.
> 
> Hi, I'm sorry, I've missed your reply.
> 
> I'm not convinced replacing bash with ash/ksh or POSIX sh will improve performance (here shell scripts aren't really executing mission-critical code that requires fastest execution) or maintainability (replacing "simple bashisms" with evals or other less readable constructs).
> What I'd agree with it, is if this change was required by licensing issues of Bash, but I think it's not the case? Or because of utilizing some predefined system/container img that for some reason (probably size, or licensing?) doesn't contain bash.
> Otherwise for me it's actually reducing maintainability.
> 
> Could you please relate?

Sure
I agree that these structure can be less readable and that evals should be handled carefully.
I am open to any other suggestions here and I will also be fine if we decide to keep bash for some scripts because of the complexity of migrating them.

Though, I still think it is a good idea to avoid bash or any other shell adherency when possible . .

About performance, this is not only my opinion but also the main shared one among Linux distributions as shown in the URLs of the first commit.
It is already demonstrated that dash is better performing than bash in most situations. I don't know what would be the impact in ONAP but if the performance increase is not viewable on one container, we might consider it over the whole number of containers and the number of times the scripts are run.

About maintaniblity, scripts with /bin/sh shebangs or without shebangs should be POSIX compliant, since /bin/sh can point by default to different shells that depends on the distribution.
If not, we will bump into troubles if we have to change the distribution or if the distribution used change its default shell.

I also think that portability in general, and more specifically scripts POSIX portability is a good thing.
Let's assume that we have a security issue similar to shellshock that affects only bash.
https://en.wikipedia.org/wiki/Shellshock_(software_bug)
With a small bash footprint, the attack surface is narrowed and if bash was the default /bin/sh shell, with POSIX compliant shells, we just need to change the default shell to annihilate the risk.
If it is not the case, in a production environment, the time the scripts are adapted, it will be probably too late.

Hope this helps

Patch-set: 8
Reviewer: Gerrit User 3685 <3685@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>