Update patch set 11
Patch Set 11: Code-Review+1

Hi,
gate is good and it seems not to break when not using keycloak.
I've got one small remark regarding the code, the scheme for keycloak is fixed to http.
I think it would worth it (in another patch or this one as you prefer, but having "green" gate is quite complicated these times...) to allow to choose between http and https

I'm not putting +2 as I'd like to have krysztof review on this one but for me it's OK to merge.

Patch-set: 11
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>
Label: Code-Review=+1