Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> Would suggest following order for review/merge on template updates
> https://gerrit.onap.org/r/c/oom/+/120618
> https://gerrit.onap.org/r/c/oom/+/120211
> and then this patch (https://gerrit.onap.org/r/c/oom/+/120756)

Vijay (Perhaps Lucas as well):
I think, we shall be fine to preserve such an order as You propose.
However, the team noticed something like this for "envsubst" container:
        volumeMounts:
        - mountPath: /config-input
          name: app-config-input
        - mountPath: /config
          name: app-config
        image: {{ include "repositoryGenerator.image.envsubst}}

The team noticed, that there is an operation done on the app-config-input configMap --> The contents are processed as ENV variables and then stored to app-config volume; then this volume is exposed to other containers.

I would assume, that we`d like to expose to the DCAE collectors the configMap (so app-config-input) itself, and not really the processed files, stored then in a app-config volume.

Could You please confirm this?

Patch-set: 4
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>