Update patch set 2
Patch Set 2:

Hello, as soon as jobbuilder is +1 (line too long is fixed) I’m OK to merge. 
Although I’ve got a question for the « future »: is it simple to remove all AAF stuff in Clamp as we all know that we want to remove it from ONAP as it’s not maintained anymore?

Patch-set: 2
Reviewer: Gerrit User 2319 <2319@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>