Update patch set 5
Patch Set 5:

> > The configuration that's pushed into Consul is the processed file, not the original configMap.
> 
> Ok, understood, but...
> This will mean, we cannot support the live configuration update, as the changes done in the configMap will not be reflected in the processed file (as far as I see, envsubst is an initContainer, not a sidecar container).
> 
> Your view? Vijay?
> Would it be fine to start with this approach, that we start with configuration load at start-up?
> Or shall we think of sth like using secret names in the configMap, so we can resolve them in the CBS-Client?

I believe there's a "best practice" recommendation somewhere (I can't find the reference) for applications to store credentials separately from other configuration information--in a separate file perhaps, or in environment variables.   If we were to do that, then we would not need to do the envsubst on the non-credential configuration, and we could mount the configMap with the non-credential configuration directly on the container.  This requires a change to the applications, so that they no longer looks for credentials in the same place as the other configuration.  Potentially this could all be handled in the CBS client, so that the main application code doesn't need to change.

Patch-set: 5
Reviewer: Gerrit User 511 <511@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>