Update patch set 9
Patch Set 9:

> Patch Set 9: Code-Review-1
> 
> (1 comment)
> 
> Just one minor catch around ifs but I think we are good apart from that

I'm not going to debate this issue, because this submission has been in review for far too long and we need to get it moving.  I'm just going to make the change without trying to defend what I did in the code.  (But it may be a while before I submit the change.  See item (2) below.)

But I will make some general remarks.

In the absence of written style/best practice guides for OOM, I believe:

(1) Working code that does not have serious deficiencies and does not deviate wildly from common         practice in OOM should be accepted, even if it isn't 100% consistent with a particular committer's opinion.
(2) Issues like this one should be identified in the first pass of a review, not in a late patchset.
Each iteration is too expensive, especially since responsible developers will regression test the change thoroughly in their own environments before submitting it--a process that can take several hours even for a small change.  And then there's gating, which can take hours, and often requires repetition over a period of several days.  It may seem to the committer like "one minor catch", but it's often a lot of work for the developer and will add days to the schedule.  If an issue really is minor and could have been caught in an earlier review, the submission should be accepted.

Post I-release (or, even better, as part of the I release), the Helm charts associated with a particular application or common code that's shared among a class of applications (such as DCAE microservices) should be the responsibility of the team developing the applications and managed by that team.   Having a giant, monolithic tree of charts with long build and test times and a small number of committers doesn't allow for efficient development and testing.  In particular, dcaegen2-services--a collection of small applications that are mostly deployed on demand after ONAP is deployed--doesn't belong in the OOM tree.

Patch-set: 9