Update patch set 6
Patch Set 6:

> Patch Set 4:
> 
> > Patch Set 4: Code-Review-1
> > 
> > hmmm... correct me if I'm wrong but shouldn't the prometheus metrics be explicitly enabled from the helm chart rather than hardcoding their enablement in the image?
> 
> This just exposes end-point with metrics (and this requires compilation-time changes). However actual metrics won't be gathered until Prometheus integration is enabled via Helm values as you mentioned. This was agreed with Vijay and discussed on OOM weekly as well.

Yeah that's true and as you remember I raised the point that those metrics should not compromise the security and I'm afraid that they do in this case.

So based on my understanding (please correct me if I'm wrong) you have a service of type NodePort here and since this patch it will be exposing to the outside of the cluster not only its main functionality but also this unprotected endpoint with prometheus metrics no?

Patch-set: 6
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>