Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> Why are 3gpp SA91-Rel16 files being added?  Is there a corresponding REQ/EPIC requiring/tracking this additions for Istanbul? This list will need to be reviewed closer by VES/VNFREQ Team and only files relevant from ONAP default processing must be included.

This a new release of 3GPP, which solves a number of bugs, which were present in previous versions.
This is still rel-16, so no new functionality is expected here, only BugFixes.

With this release, 3GPP has actually restructured the files - e.g. there are new "comDefs.yaml" files which contain definitions of common objects used in different management services.
This is why the amount of files is different than previously, but essentially the functionality is the same.

We do not have an intention to support more, than what was present previously in Honolulu release.
It is simply, that 3GP is restructuring the files, to simplify the definitions and make them more manageable. 

Still, it is absolutely fine to review the files, and if there are any files, which shall not be present (I think, I`ll ask about some of them), then we shall remove them.

Patch-set: 1
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>