Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
> 
> (2 comments)
> 
> > Patch Set 1: Code-Review-1
> > 
> > Ok I believe we need to discuss this during oom meeting together with Sylvain.
> > For me there is no point in providing additional abstraction layer for every part of deployment template. I'd rather allow to put a piece of yaml in values for that
> 
> -1 on few inline comments noted.
> 
> Hey Krzysztof - The usecase here is primarily for certain DCAE Microservice requiring Configmap to be mounted into application; we have this capability supported under Cloudify/Blueprint deployment (where same parameters - configmapname, mountpath) are provided as input and K8s plugin setup the volumemount). As all DCAE services will be using dcaegen2-services-common.microserviceDeployment template, hence need this change to support in generic manner under helm migration.
> 
> Btw, not sure what you refer by including yaml directly in values.yaml? Would you have an example?

Yeah sure. Just to be 100% clear I'm fine with the functionality that you are trying to achieve here there is no issues with that. I've just concerns about the way of achieving it

We had a similar discussion around secrets. We had two approaches:
1) Define an "abstraction layer" for microservices to add environment variables
2) Allow microservices to inject a pieces of yaml into their values.yaml that would be then directly copied to the right place in the deployment.yaml

With envs we decided to go with #1 as it was your team preference but as this gets popping up here and there I'd like to have this discussion again

Patch-set: 1
Reviewer: Gerrit User 4965 <4965@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>