Update patch set 19
Patch Set 19:

on the last gate (19 on Azure 3) most of the errors are due to the fact that the models we distribute in the different tests are not properly distributed. the SO never gets them.

No link with the patch (gate 18 was correct)

Patch-set: 19
Reviewer: Gerrit User 1215 <1215@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>