[PMS] Persistent storage of policies and type definitions
Changed the deployment to a stateful set.

Even Deployments with one replica using ReadWriteOnce volume
are not recommended. This is because the default Deployment
strategy creates a second Pod before bringing down the first Pod
on a recreate. The Deployment may fail in deadlock as the
second Pod can't start because the ReadWriteOnce volume is
already in use, and the first Pod won't be removed because
the second Pod has not yet started.
Instead, use a StatefulSet with ReadWriteOnce volumes.

Issue-ID: CCSDK-3256
Signed-off-by: PatrikBuhr <patrik.buhr@est.tech>
Change-Id: Ie3906cdc95481f14fe8c2f15aa7df1b5c848df34