Enable apex-pdp to mount cert files during deployement
This changeset allows APEX-PDP charts to mount keysstore/truststore files into the
container during deployment.
Note: This is not for the ONAP common certificates, but to allow the
possibility to add custom certificates specific to deployments.
With this change, adding the required certificate files into
resources/certs folder will mount it into the respective location in the
container, which in turn can be used by apex-pdp to connect to different
systems specific to the setup.

Change-Id: Id5edcd998e34a17bf93c4e592fa45737c9e60d35
Issue-ID: POLICY-3232
Signed-off-by: a.sreekumar <ajith.sreekumar@bell.ca>