[COMMON][DCAE] Allow internal only ports
Sometimes, we need to generate a service for ports that are internal
only. So, when the service is "NodePort", it's an issue.
By adding a toggle in service template + marking the ports as internal
only, we'll now create two services (one internal, one external) if
needed.
Adding the option in Ves Collector service.

Issue-ID: OOM-1
Signed-off-by: Sylvain Desbureaux <sylvain.desbureaux@orange.com>
Change-Id: I27cf48f79dddc849133c27527531efb66cec4162