Update patch set 7
Patch Set 7:

> Patch Set 7:
> 
> > Patch Set 7:
> > 
> > any idea why dcae dashboard is constanly failing and how to fix it?
> 
> I am fine with the idea to merge
> it would be better however than the HC of the component that is directly impacted would be PASS
> 
> I created a JIRA to track the regular DCAE HC error
> https://jira.onap.org/browse/DCAEGEN2-2800
> I know that there are some refactoring planned, if it makes sense to amend or disable the HC, please amend the test.
> HC are the basic tests to give feedback on a component
> Ideally we should rely on this test to perform or not E2E tests

Yes, I agree, it would make sense to have the DCAE HC up-and-runing.
However that is not something we could do at the moment as a team (we haven`t been working on that part of DCAE in the past). Generally, the cfgMaps, we`re updating here are currently used in DCAE ves-openapi-manager only. There is no chance, that they have impact on DCAE HC...

Patch-set: 7
Reviewer: Gerrit User 5597 <5597@14b4e41f-c4e3-4fb9-9955-6a9b5656895a>