Update patch set 6
Patch Set 6:

same behavior than a recent patch on AAI

the smoke tests basic_vm and basic_vm_macro are failed, it seems to be linked to the deletion of ONAP resources
onboarding, instantiation, VM creation, VM deletion is failed
but something goes wrong when trying to delete ONAP associated resources 

2021-06-15T23:35:10.662Z|cd59e4e6-17f0-434b-99f3-7b59eed0b630|http-nio-8080-exec-5|onap/so/infra/orchestrationRequests/v7/0873bd96-c03b-4d87-bea2-58e04e9c5a8b|InfraPortalClient|Unknown-Target-Entity|http://so-request-db-adapter.onap:8083/infraActiveRequests/0873bd96-c03b-4d87-bea2-58e04e9c5a8b|DEBUG|500||Response body: {
  "requestId" : "0873bd96-c03b-4d87-bea2-58e04e9c5a8b",
  "requestStatus" : "IN_PROGRESS",
  "resourceStatusMessage" : "The vf module was successfully deleted in the cloud",
  "statusMessage" : "Exception occured during vnf adapter: org.onap.so.openstack.utils.StackCreationException: Stack Creation Failed Openstack Status: CREATE_FAILED Status Reason: Resource CREATE failed: ResourceInError: resources.Ubuntu20_server_0: Went to status ERROR due to \"Message: No valid host was found. , Code: 500\". Stack rollback suppressed, stack not deleted. The resource was rollbacked in openstack: true.",

https://logs.onap.org/onap-integration/gating//121624-6/k8s-teardown/k8s-teardown/kubernetes-status/pod-onap-so-74ff9c9699-hmpgk-so.log

note no valid host may mean that there is no more resources in the gating lab, then the nova scheduler cannot find a good host to run the test. Maybe old (not cleaned) VMs are remaining in the lab leading to this error.

Patch-set: 6