[ONAP] Disable some multicloud drivers at installation
Several multicloud drivers are installed by default.
The code of most of these drivers is unchanged since several versions.
It has been suggested to save resource to keep only the mandatory opensource
driver and consider the other drivers as possible addons.

This patch disables this driver from the default installation.

Issue-ID: MULTICLOUD-1362

Signed-off-by: morganrol <morgan.richomme@orange.com>
Change-Id: I2a1e42241b4b3dfd5bb99a741360aec6fa36e331