Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> we are lack of commitment to support multicloud-pike driver, I prefer keeping multicloud-starlingx, disabling multicloud-pike.

Hi Bin

If I am not wrong, none of the drivers (pike, starlingx, vio, windriver) has changed since frankfurt
		                  H	 G MC     G 	  F
multicloud	                1.6.0	1.6.0	1.6.0	1.5.1
multicloud-azure	        None    1.2.4	1.2.4	1.2.4
multicloud-fcaps	        1.5.5	1.5.5	1.5.5	1.5.5
multicloud-k8s	                0.8.1	0.7.0	0.7.0	0.6.0
multicloud-k8s-mongo	        4.0.8	4.0.8	4.0.8	4.0.8
multicloud-pike	                1.5.5	1.5.5	1.5.5	1.5.5
multicloud-starlingx	        1.5.5	1.5.5	1.5.5	1.5.5
multicloud-vio	                1.4.1	1.4.1	1.4.1	1.4.1
multicloud-windriver	        1.5.5	1.5.5	1.5.5	1.5.5

the multicloud-pike is used in tests integrated in CI since guilin but as far as I know there is no starlingx resources available in CI for tests so there is no tests dealing with starlingx driver at the moment (except the driver tests healthcheck tests).
It could be a good opportunity to create a CI chain with a starlingx target, we could probably have done that on the intel/windriver lab but it is not really doable at the moment.

So my recommendation would be to keep the driver pike as it is currently used in community tests.

Moreover disabling the drivers does not mean removing them, they still can be installed if needed if associated cloud resources are used.
So if companies are using the resources requiring specific drivers in CI they can easily override the installation.

Patch-set: 1