Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > dcae policy handler is failing, is this related?
> 
> Yeah, same issue as commented on https://gerrit.onap.org/r/c/oom/+/121733; PH failure though is unrelated to this change.
> 
> Did anything change recently on how PV/volume should be created? The PH pod is stuck on waiting.

It's hard to tell from the error message what volume or what volume mount is causing the problem here.  The filebeat container is running, so the filebeat container mounts must be OK.  The policy handler container has 2 volume mounts: the log volume (shared with filebeat) at "/app/logs" and the volume holding TLS certificate materials at "/app/etc/tls/certs/".  The associated volumes are emptyDir volumes.  The error is complaining that the destination of the mount is "logs" and isn't an absolute path.

We see the same error in the daily master run from 2021-06-24, which is the last one I can find:
https://logs.onap.org/onap-integration/daily/onap_daily_pod4_master/2021-06/24_06-08/infrastructure-healthcheck/k8s/kubernetes-status/pod-onap-dcae-policy-handler-9d89bdc5-csgks.html

There was a change back in March, changing the mount path for the volumes mounts from
/opt/app/policy_handler/logs to /app/logs
and
/opt/app/policy_handler/etc/tls/certs/ to /app/etc/tls/certs/
along with a change to the policy handler version.
Seems too long ago to be making a difference now.

Patch-set: 11