Update patch set 12
Patch Set 12:

> Patch Set 12:
> 
> Sylvain/Krzysztof - The gating failures are un-related; as this is new MS introduction. Are we good to merge this change yet?

@Vijay shall we do something for the DCAE Healthcheck, it is very regularly failing
I created a JIRA some weeks ago https://jira.onap.org/browse/DCAEGEN2-2800
I know that you tried to tune the resources for the DCAE dashboard but I am not sure it is enough

Shall we temporarily disable it? how could we hlep consolidating this healthcheck.
Ideally we would like to vote with a onap-k8s and onap-helm OK + 100% PASS on HC and 100% on smoke
if some HC are systematically failing, it will be hard to automate the vote.

Patch-set: 12