Update patch set 11
Patch Set 11:

> Patch Set 11:
> 
> > Patch Set 11:
> > 
> > > Patch Set 11:
> > > 
> > > dcae policy handler is failing, is this related?
> > 
> > Yeah, same issue as commented on https://gerrit.onap.org/r/c/oom/+/121733; PH failure though is unrelated to this change.
> > 
> > Did anything change recently on how PV/volume should be created? The PH pod is stuck on waiting.
> 
> It's hard to tell from the error message what volume or what volume mount is causing the problem here.  The filebeat container is running, so the filebeat container mounts must be OK.  The policy handler container has 2 volume mounts: the log volume (shared with filebeat) at "/app/logs" and the volume holding TLS certificate materials at "/app/etc/tls/certs/".  The associated volumes are emptyDir volumes.  The error is complaining that the destination of the mount is "logs" and isn't an absolute path.
> 
> We see the same error in the daily master run from 2021-06-24, which is the last one I can find:
> https://logs.onap.org/onap-integration/daily/onap_daily_pod4_master/2021-06/24_06-08/infrastructure-healthcheck/k8s/kubernetes-status/pod-onap-dcae-policy-handler-9d89bdc5-csgks.html
> 
> There was a change back in March, changing the mount path for the volumes mounts from
> /opt/app/policy_handler/logs to /app/logs
> and
> /opt/app/policy_handler/etc/tls/certs/ to /app/etc/tls/certs/
> along with a change to the policy handler version.
> Seems too long ago to be making a difference now.

In my local environment, I rebased against the latest master (as of yesterday late afternoon East Coast USA time), rebuilt the charts, and redeployed an ONAP subset with DCAE.  The policy handler came up properly.  (If I had seen the problem, I would have had access to the underlying docker volume as well as to the rendered charts, which might shed some light on the problem.)

I notice that today's daily integration run using the master branch is still showing the problem.
https://logs.onap.org/onap-integration/daily/onap_daily_pod4_master/2021-06/29_07-36/infrastructure-healthcheck/k8s/kubernetes-status/pod-onap-dcae-policy-handler-9d89bdc5-pbf59.html

It would appear that this has something to do with the environment where gating and integration are running, or it is something that happens only when something more than my ONAP subset is deployed.

Is there any possibility of getting access to the following?
  -- The rendered charts that are driving the gating and integration runs.  I'm interested in these because the error message in the logs complains that the mount "destination" is not absolute.  I'm wondering if somehow the value in the charts "/app/logs" is being changed to "logs" somehow.
  -- The contents of the /var/lib/docker/volumes/ subdirectory that the error message identifies as the "source" of the mount.  This might give a clue about what content cannot be mounted.

It's pretty clear that this change is not responsible for the behavior we're seeing.  The problem is already present in the master branch.  We need to find out what's happening.

Patch-set: 11