Update patch set 11
Patch Set 11:

> I was able to reproduce the Policy-handler volume creation error on one of WR setup - shared the info with Jack, he is troubleshooting there.  
> 
> Btw Policy-Handler problem started manifesting since 05/26/21 on both master and honolulu daily runs - looking back into OOM commits for that period - did not find anything that could attribute to PV setup failure hence this could be env thing (possibly related to docker version itself). Anyways, we are checking this. As the issue is outside of the majority of DCAE patches - vote +1 to move forward on building up DCAE queue of submission.

I looked at the environment where Vijay has reproduced the problem.  I didn't see any good clues.  The charts are properly rendered.  The /var/lib/docker/.... directory (the "source" of the mount) is empty, which is to be expected for an emptyDir volume.   I tried editing the k8s Deployment to change the mountPath from "/app/logs" to "/app/logs/", but this did not get rid of the problem.  (I'm still not entirely certain what mount the error message is talking about.  There are two volume mounts on the dcae-policy-handler container, one for the logs and one for the TLS certificate info.  I'm guessing, based on the name "logs" in the error message, that it's the log volume mount.  But I could be wrong.)

I don't think PVs/PVCs are an issue here, since we're only using emptyDir volumes on the policy handler container.

I did not that the environment where Vijay is running and the environment where I ran my tests are running different versions of containerd (1.4.4 in Vijay's environment, 1.3.3 in mine) and runc (1.0.0~rc95 in Vijay's environment, 1.0.2-dev in mine).

Patch-set: 11