Update patch set 11
Patch Set 11:

On further investigation, I have found that the "logs" volume mentioned in the error message is *not* one of the two volume mounts explicitly mentioned in the Helm chart for the policy handler deployment.  It's a spurious volume defined in the Dockerfile:
VOLUME logs
It's present in Vijay's environment and it's present in mine.  But apparently my environment can handle the relative path and Vijay's (and gating) cannot.  As far as I can tell, this volume is not needed for anything, and we should drop it from the Dockerfile.

[How to see this:  First, do a kubectl describe pod and get the Container ID for the container.  (You don't need the docker:// part, just the first few characters of the hex string).  Then do a docker inspect -f {{ .Mounts }} to see all of the mounts on the container.
You can also look at the docker image with a docker inspect, and you'll see the "logs" volume from the Dockerfile VOLUME definition>]

Patch-set: 11